refactor: split parser mappers into modules
